pytest: Separate formatter tests
