+++ server/camcops_server/cc_modules/cc_formatter.py
@@ -30,7 +30,6 @@ camcops_server/cc_modules/cc_formatter.py
 """
 
 from string import Formatter
-from unittest import TestCase
 
 from typing import Any, Mapping, Sequence, Tuple
 
@@ -112,32 +111,3 @@ class SafeFormatter(Formatter):
         test_dict = {k: "" for k in self._allowed_keys}
 
         self.format(format_string, **test_dict)
-
-
-class SafeFormatterTests(TestCase):
-    def setUp(self) -> None:
-        super().setUp()
-
-        self.formatter = SafeFormatter(["forename", "surname", "email"])
-
-    def test_formats_with_allowed_keys(self) -> None:
-        output = self.formatter.format("{forename} {surname} <{email}>",
-                                       forename="Erin", surname="Byrne",
-                                       email="erin.byrne@example.com")
-        self.assertEqual(output, "Erin Byrne <erin.byrne@example.com>")
-
-    def test_format_raises_with_disallowed_keys(self) -> None:
-        with self.assertRaises(KeyError):
-            self.formatter.format("{email.__class__}")
-
-    def test_returns_valid_parameters_string(self) -> None:
-        self.assertEqual(self.formatter.get_valid_parameters_string(),
-                         "{forename}, {surname}, {email}")
-
-    def test_validate_raises_key_error_for_unknown_key(self) -> None:
-        with self.assertRaises(KeyError):
-            self.formatter.validate("{phone}")
-
-    def test_validate_raises_value_error_for_mismatched_brackets(self) -> None:
-        with self.assertRaises(ValueError):
-            self.formatter.validate("{forename")
